Work on the Gas Mixing Model and attempt at creating an example app.
The gas mixing model was worked on, there is currently a working model for two gasses at equilibrium.
More work needs to be done to create initial gas model as well as the mathematical calculator.
An attempt was made to make a sample app using Andriod studio, however some unexpected errors in relation to dependencies and Andriod API version came up.