Paramagnet Simulation Nearly Finished
Big update to the paramagnet simulation, where a working calculator is shown. The function is fairly simple however works with all reasonable numbers. It can take a float input and takes in exponential powers in the form "5e6".
The graph shows arrows of two orientations (hence dipole). The proportions of which can be changed by N and n.